% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{bregress2}
\alias{bregress2}
\title{Simulate data from time-varying time series regression model with change points}
\usage{
bregress2(nn, cp = 0, delta = 0, type = "norm")
}
\arguments{
\item{nn}{sample size}

\item{cp}{number of change points. If cp is between 0 and 1, it specifies the location of the single change point}

\item{delta}{double, magnitude of the jump}

\item{type}{type of distributions of the innovations, default normal. It can also be "t4", "t5" and "t6".}
}
\value{
a list of data, x covariates, y response and e error.
n = 300
data = bregress2(n, 2, 1) # time series regression model with 2 changes points
}
\description{
Simulate data from time-varying time series regression model with change points
}
