% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\docType{data}
\name{hk_data}
\alias{hk_data}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
Fan, J., and Zhang, W. (1999). Statistical estimation in varying coefficient models. The annals of Statistics, 27(5), 1491-1518.
}
\author{
T. S. Lau
}
\keyword{data}
