% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MV_critical}
\alias{MV_critical}
\title{Statistics-adapted values for extended minimum volatility selection.}
\usage{
MV_critical(
  y,
  data,
  R,
  gridm,
  gridtau,
  type = 1L,
  cvalue = 0.1,
  B = 100L,
  lrvmethod = 1L,
  ind = 2L,
  rescale = 0L
)
}
\arguments{
\item{y, }{vector, as used in the Heter_LRV}

\item{data, }{list, a list of data}

\item{R, }{a cube of standard.normal random variables.}

\item{gridm, }{vector, a grid of candidate m's.}

\item{gridtau, }{vector, a grid of candidate tau's.}

\item{type, }{integer, 1 KPSS 2 RS 3 VS 4 KS}

\item{cvalue, }{double, 1-qunatile for the calculation of bootstrap variance, default 0.1.}

\item{B, }{integer, number of iterations for the calculation of bootstrap variance}

\item{lrvmethod, }{integer, see also Heter_LRV}

\item{ind, }{integer, the type of kernel,  see also Heter_LRV}

\item{rescale, }{bool, whether to rescale when positiveness of the matrix is not obtained. default 0}
}
\value{
matrix of critical values
}
\description{
Calulation of the variance of the bootstrap statistics for the extended minimum volatility selection.
}
\references{
#' Bai, L., and Wu, W. (2023). Detecting long-range dependence for time-varying linear models. To appear in Bernoulli
}
\seealso{
Heter_LRV
}
