% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{Qct_reg}
\alias{Qct_reg}
\title{Simulate data from time-varying time series regression model}
\usage{
Qct_reg(T_n, param, type = 1)
}
\arguments{
\item{T_n}{int, sample size}

\item{param}{list, a list of parameters}

\item{type}{type = 1 means the long memory expansion begins from its infinte past, type = 2 means the long memory expansion begins from t = 0}
}
\value{
list, a list of data, covariates, response and errors.(before and after fractional difference)
}
\description{
Simulate data from time-varying time series regression model
}
