% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Heter_LRV}
\alias{Heter_LRV}
\title{Long-run covariance matrix estimators}
\usage{
Heter_LRV(
  e,
  X,
  m,
  tau_n = 0,
  lrv_method = 1L,
  ind = 2L,
  print_deg = 0L,
  rescale = 0L
)
}
\arguments{
\item{e, }{vector, if the plug-in estimator is used, e should be the vector of residuals, OLS or nonparametric ones. If the difference-based debiased method is adopted, e should be the reponse time series, i.e., \mjseqn{y}. Specially, e should also be the reponse time series, i.e., \mjseqn{y}, if the plug-in estimator using the \mjseqn{\breve{\beta}}, the pilot estimator proposed in Bai and Wu (2023).}

\item{X, }{a matrix \mjseqn{n\times p}}

\item{m, }{integer, the window size.}

\item{tau_n, }{double, the smoothing parameter in the estimator. If tau_n is 0, a rule-of-thumb value will be automatically used.}

\item{lrv_method, }{the method of long-run variance estimation, lrvmethod = 0 uses the plug-in estimator in Zhou (2010), lrvmethod = 1 offers the debias difference-based estimator in Bai and Wu (2023), lrvmethod = 2 provides the plug-in estimator using the \mjseqn{\breve{\beta}}, the pilot estimator proposed in Bai and Wu (2023)}

\item{ind, }{types of kernels
\itemize{
\item 1 Triangular \mjseqn{1-|u|}, \mjseqn{u \le 1}
\item 2 Epanechnikov kernel \mjseqn{3/4(1-u^{2})}, \mjseqn{u \le 1}
\item 3 Quartic \mjseqn{15/16(1-u^{2})^{2}}, \mjseqn{u \le 1}
\item 4 Triweight \mjseqn{35/32(1-u^{2})^{3}}, \mjseqn{u \le 1}
\item 5 Tricube  \mjseqn{70/81(1-|u|^{3})^{3}}, \mjseqn{u \le 1}
}}

\item{print_deg, }{bool, whether to print information of non-positiveness, default 0\mjseqn{n\times p}{}}

\item{rescale, }{bool, whether to use rescaling to correct the negative eigenvalues, default 0}
}
\description{
\loadmathjax The function provides a wide range of estimators for the long-run covariance matrix estimation in non-stationary time series with covariante.

Zhou, Z. and Wu, W. B. (2010). Simultaneous inference of linear models with time varying coefficients.J. R. Stat. Soc. Ser. B. Stat. Methodol., 72(4):513–531.
}
