% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MV_ise_heter_critical}
\alias{MV_ise_heter_critical}
\title{MV method}
\usage{
MV_ise_heter_critical(critical, neighbour)
}
\arguments{
\item{critical, }{a matrix of critical values}

\item{neighbour, }{integer, number of neighbours}
}
\value{
a list of results,
\itemize{
\item minp: optimal row number
\item minq: optimal column number
\item min_ise: optimal value
}
}
\description{
Selection of smoothing parameters for bootstrap tests by choosing the index minimizing the volatility of bootstrap statistics or long-run variance estimators in the neighborhood computed before.
}
\references{
#' Bai, L., and Wu, W. (2023). Detecting long-range dependence for time-varying linear models. To appear in Bernoulli
}
