% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{Qct_reg}
\alias{Qct_reg}
\title{Simulate data from time-varying time series regression model}
\usage{
Qct_reg(T_n, param, type = 1)
}
\arguments{
\item{T_n}{int, sample size}

\item{param}{list, a list of parameters}

\item{type}{type = 1 means the long memory expansion begins from its infinite past, type = 2 means the long memory expansion begins from t = 0}
}
\value{
list, a list of data, covariates, response and errors.(before and after fractional difference)
}
\description{
Simulate data from time-varying time series regression model
}
\examples{
param = list(d = -0.2, heter = 2, tvd = 0,
tw = 0.8, rate = 0.1, cur = 1, center = 0.3,
ma_rate =  0, cov_tw =  0.2, cov_rate = 0.1,
cov_center = 0.1, all_tw  = 1, cov_trend = 0.7)
n = 500
data = Qct_reg(n, param)
}
