% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{Qt_data}
\alias{Qt_data}
\title{Simulate data from time-varying trend model}
\usage{
Qt_data(T_n, param)
}
\arguments{
\item{T_n}{integer, sample size}

\item{param}{a list of parameters
\itemize{
\item tw double, squared root of  variance  of the innovations
\item rate double, magnitude of non-stationarity
\item center double, the center of the ar coefficient
\item ma_rate double, ma coefficient
}}
}
\value{
a vector of non-stationary time series
}
\description{
Simulate data from time-varying trend model
}
