% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_T}
\alias{sim_T}
\title{bootstrap distribution}
\usage{
sim_T(X, t, sigma, m, B, type = 0L)
}
\arguments{
\item{X, }{matrix of covariates}

\item{t, }{vector of time points}

\item{sigma, }{a cube of long-run covariance function.}

\item{m, }{int value of window size}

\item{B, }{int, number of iteration}

\item{type}{type of tests, residual-based or coefficient-based}
}
\value{
a vector of bootstrap statistic
}
\description{
bootstrap distribution of the gradient based structural stability test
}
